Type Definition kusama_runtime::AllPalletsWithSystemReversed

pub type AllPalletsWithSystemReversed = (MessageQueue, XcmPallet, StateTrieMigration, Crowdloan, Auctions, Slots, Registrar, ParaAssignmentProvider, ParasSlashing, ParasDisputes, ParaSessionInfo, Hrmp, Dmp, Initializer, Paras, ParaScheduler, ParaInherent, ParaInclusion, ParasShared, Configuration, ParachainsOrigin, FastUnstake, NominationPools, VoterList, NisCounterpartBalances, Nis, ElectionProviderMultiPhase, ChildBounties, Bounties, Preimage, Multisig, Proxy, Scheduler, Vesting, Recovery, Society, Identity, Utility, Claims, Whitelist, Origins, FellowshipReferenda, FellowshipCollective, Referenda, ConvictionVoting, Treasury, AuthorityDiscovery, ImOnline, Grandpa, Session, BeefyMmrLeaf, Mmr, Beefy, Historical, Offences, Staking, Authorship, TransactionPayment, Balances, Indices, Timestamp, Babe, System);

👎Deprecated: Using reverse pallet orders is deprecated. use only AllPalletsWithSystem or AllPalletsWithoutSystem

All pallets included in the runtime as a nested tuple of types in reversed order.