Module kusama_runtime::xcm_config

XCM configurations for the Kusama runtime.

Structs

Ancestry

Our XCM location ancestry - i.e. what, if anything, Parent means evaluated in our context. Since Kusama is a top-level relay-chain, there is no ancestry.

BaseXcmWeight

The amount of weight an XCM operation takes. This is a safe overestimate.

CheckAccount

The check account, which holds any native assets that have been teleported out and not back in (yet).

CouncilBodyId

KsmLocation

The location of the KSM token, from the context of this chain. Since this token is native to this chain, we make it synonymous with it and thus it is the Here location, which means “equivalent to the context”.

Kusama

KusamaForEncointer

KusamaForStatemine

KusamaNetwork

The Kusama network ID. This is named.

MaxInstructions

Maximum number of instructions in a single XCM fragment. A sanity check against weight calculations getting too crazy.

OnlyParachains

XcmConfig

Type Definitions

Barrier

The barriers one of which must be passed for an XCM message to be executed.

LocalAssetTransactor

Our asset transactor. This is what allows us to interest with the runtime facilities from the point of view of XCM-only concepts like MultiLocation and MultiAsset.

LocalOriginToLocation

Type to convert an Origin type value into a MultiLocation value which represents an interior location of this chain.

SovereignAccountOf

The canonical means of converting a MultiLocation into an AccountId, used when we want to determine the sovereign account controlled by a location.

TrustedTeleporters

XcmRouter

The XCM router. When we want to send an XCM message, we use this type. It amalgamates all of our individual routers.